requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml>=6.0
wandb>=0.15.0
